Using display name for the LTI POST parameters.
Some LTI providers respond differently to the name of the org used in the context_id. For multi-tenancy purposes the org must be diferent for each tenant, but the LTI should still behave the same. In this cases the block now allows the user to send the Display name of the organization instead